correlation: move 'map' submodule's contents into mod.rs
The original code produced a warning.
